Change variationId in requests to long

+++ src/main/kotlin/com/ecwid/apiclient/v3/dto/variation/request/AdjustVariationInventoryRequest.kt
@@ -6,7 +6,7 @@ import com.ecwid.apiclient.v3.impl.RequestInfo
 
 data class AdjustVariationInventoryRequest(
 	val productId: Int = 0,
-	val variationId: Int = 0,
+	val variationId: Long = 0,
 	val checkLowStockNotification: Boolean = false,
 	val quantityDelta: Int = 0
 ) : ApiRequest {
